refactor: remove unused files, move for useful hierarchy
