--- src/loaders/flattened/__init__.py
@@ -1,3 +1,0 @@
-"""Loaders for importing from disk."""
-
-from .local_feature_loaders import *  # noqa
