+++ src/timeseriesflattener/testing/load_synth_data.py
