+++ src/timeseriesflattener/testing/utils_for_testing.py
@@ -10,7 +10,7 @@ import pytest
 from pandas import DataFrame
 from pandas.testing import assert_series_equal
 
-from loaders.synth.raw.load_synth_data import (
+from timeseriesflattener.testing.load_synth_data import (
     load_synth_outcome,
     load_synth_prediction_times,
 )
