--- tests/test_data/test_hf/test_hf_embeddings.py
@@ -1,8 +1,0 @@
-"""Tests for generating huggingface embeddings."""
-
-_test_hf_embeddings = [i for i in range(384)]
-
-
-TEST_HF_EMBEDDINGS = [
-    "embedding-" + str(dimension) for dimension in _test_hf_embeddings
-]
