+++ tests/test_timeseriesflattener/test_errors.py
@@ -4,7 +4,7 @@ import pytest
 
 from timeseriesflattener.feature_spec_objects import PredictorSpec
 from timeseriesflattener.flattened_dataset import FlattenedDataset
-from timeseriesflattener.utils_for_testing import (
+from timeseriesflattener.testing.utils_for_testing import (
     str_to_df,  # pylint: disable=import-error
 )
