+++ tests/test_timeseriesflattener/test_feature_spec_objects.py
@@ -2,7 +2,7 @@
 import pandas as pd
 import pytest
 
-from loaders.synth.raw.load_synth_data import (  # pylint: disable=unused-import
+from timeseriesflattener.testing.load_synth_data import (  # pylint: disable=unused-import
     synth_predictor_float,
 )
 from timeseriesflattener.feature_spec_objects import AnySpec
